- Fix msvc build
SVN r2321 (0.6-dev)

+++ common/p_setup.cpp
@@ -1458,7 +1458,7 @@ void P_SetupLevel (char *lumpname, int position)
 		// [SL] 2011-07-01 - Check to see if the reject table is of the proper size
 		// If it's too short, the reject table should be ignored when 
 		// calling P_CheckSight
-		if (W_LumpLength(lumpnum + ML_REJECT) < ((unsigned int)ceil(numsectors * numsectors / 8)))
+		if (W_LumpLength(lumpnum + ML_REJECT) < ((unsigned int)ceil((float)(numsectors * numsectors / 8))))
 		{
 			Printf(PRINT_HIGH, "Reject matrix is not valid and will be ignored.\n");
 			rejectempty = true;
